# virtual-queue-back/src/main/java/mx/edu/uteq/virtual_queue_back/auth/dto/UserSummaryDTO.java
package mx.edu.uteq.virtual_queue_back.auth.dto;

import java.util.UUID;

import mx.edu.uteq.virtual_queue_back.common.UserRole;

public record UserSummaryDTO(
		UUID id,
		String username,
		String fullName,
		UserRole role,
		UUID placeId,
		String placeName) {
}